+++ cli.js
@@ -1,17 +1,54 @@
 #!/usr/bin/env node
 const path = require('path')
 const sao = require('sao')
+const cac = require('cac')
+const chalk = require('chalk')
+const envinfo = require('envinfo')
+const { version } = require('./package.json')
 
 const generator = path.resolve(__dirname, './')
-// In a custom directory or current directory
-const outDir = path.resolve(process.argv[2] || '.')
 
-console.log(`> Generating Nuxt.js project in ${outDir}`)
+const cli = cac('create-nuxt-app')
 
-// See https://sao.js.org/#/advanced/standalone-cli
-sao({ generator, outDir, logLevel: 2 })
-  .run()
-  .catch((err) => {
-    console.trace(err)
-    process.exit(1)
+const showEnvInfo = async () => {
+  console.log(chalk.bold('\nEnvironment Info:'))
+  const result = await envinfo
+    .run({
+      System: ['OS', 'CPU'],
+      Binaries: ['Node', 'Yarn', 'npm'],
+      Browsers: ['Chrome', 'Edge', 'Firefox', 'Safari'],
+      npmGlobalPackages: ['nuxt', 'create-nuxt-app']
+    })
+  console.log(result)
+  process.exit(1)
+}
+
+cli
+  .command('[out-dir]', 'Generate in a custom directory or current directory')
+  .option('-e, --edge', 'To install `nuxt-edge` instead of `nuxt`')
+  .option('-i, --info', 'Print out debugging information relating to the local environment')
+  .option('--verbose', 'Show debug logs')
+  .action((outDir = '.', cliOptions) => {
+    if (cliOptions.info) {
+      return showEnvInfo()
+    }
+    console.log()
+    console.log(chalk`{cyan create-nuxt-app v${version}}`)
+    console.log(chalk`✨  Generating Nuxt.js project in {cyan ${outDir}}`)
+
+    const { verbose } = cliOptions
+    const logLevel = verbose ? 4 : 2
+    // See https://saojs.org/api.html#standalone-cli
+    sao({ generator, outDir, logLevel, cliOptions })
+      .run()
+      .catch((err) => {
+        console.trace(err)
+        process.exit(1)
+      })
   })
+
+cli.help()
+
+cli.version(version)
+
+cli.parse()