+++ template/frameworks/adonis/providers/NuxtProvider.js
@@ -1,20 +1,23 @@
 const { ServiceProvider } = require('@adonisjs/fold')
-const { Nuxt, Builder } = require('nuxt<% if (edge) { %>-edge<% } %>')
+const { Nuxt, Builder } = require('nuxt<%= edge %>')
 
 class NuxtProvider extends ServiceProvider {
-  register() {
+  register () {
     this.app.singleton('Service/Nuxt', () => {
       const config = this.app.use('Config').get('nuxt')
       return new Nuxt(config)
     })
   }
 
-  async boot() {
-    const nuxt = this.app.use('Service/Nuxt')
-    if (nuxt.options.dev) {
-      await new Builder(nuxt).build()
-    } else {
-      await nuxt.ready()
+  async boot () {
+    const Helpers = this.app.use('Helpers')
+    if (!Helpers.isAceCommand()) {
+      const nuxt = this.app.use('Service/Nuxt')
+      if (nuxt.options.dev) {
+        await new Builder(nuxt).build()
+      } else {
+        await nuxt.ready()
+      }
     }
   }
 }