+++ template/frameworks/ava/test/helpers/setup.js
@@ -1,8 +1,0 @@
-require('browser-env')()
-const hooks = require('require-extension-hooks')
-const Vue = require('vue')
-
-Vue.config.productionTip = false
-
-hooks('vue').plugin('vue').push();
-hooks(['vue', 'js']).plugin('babel').push();