+++ template/frameworks/ava/test/specs/Logo.spec.js
@@ -2,7 +2,7 @@ import { mount } from '@vue/test-utils'
 import test from 'ava'
 import Logo from '../../components/Logo.vue'
 
-test('is a Vue instance', t => {
+test('is a Vue instance', (t) => {
   const wrapper = mount(Logo)
   t.is(wrapper.isVueInstance(), true)
 })