+++ template/frameworks/buefy/pages/index.vue
@@ -1,26 +1,33 @@
 <template>
   <section class="section">
     <div class="columns is-mobile">
-
       <card
         title="Free"
         icon="github-circle"
       >
-        Open source on <a href="https://github.com/buefy/buefy"> GitHub</a>
+        Open source on <a href="https://github.com/buefy/buefy">
+          GitHub
+        </a>
       </card>
 
       <card
         title="Responsive"
         icon="cellphone-link"
       >
-        <b class="has-text-grey">Every</b> component is responsive
+        <b class="has-text-grey">
+          Every
+        </b> component is responsive
       </card>
 
       <card
         title="Modern"
         icon="alert-decagram"
       >
-        Built with <a href="https://vuejs.org/">Vue.js</a> and <a href="http://bulma.io/">Bulma</a>
+        Built with <a href="https://vuejs.org/">
+          Vue.js
+        </a> and <a href="http://bulma.io/">
+          Bulma
+        </a>
       </card>
 
       <card
@@ -29,7 +36,6 @@
       >
         No other internal dependency
       </card>
-
     </div>
   </section>
 </template>