+++ template/frameworks/hapi/server/index.js
@@ -1,24 +1,26 @@
-const Hapi = require('hapi')
 const consola = require('consola')
-const HapiNuxt = require('hapi-nuxt')
+const Hapi = require('@hapi/hapi')
+const HapiNuxt = require('@nuxtjs/hapi')
 
-const server = new Hapi.Server({
-  host: process.env.HOST || '127.0.0.1',
-  port: process.env.PORT || 3000
-})
+async function start () {
+  const server = new Hapi.Server({
+    host: process.env.HOST || '127.0.0.1',
+    port: process.env.PORT || 3000
+  })
 
-server
-  .register({
-    plugin: HapiNuxt
+  await server.register({
+    plugin: HapiNuxt,
+    options: {}
   })
-  .then(() => server.start())
-  .then(() =>
-    consola.ready({
-      message: `Server running at: ${server.info.uri}`,
-      badge: true
-    })
-  )
-  .catch(err => {
-    consola.error(err)
-    throw err
+
+  await server.start()
+
+  consola.ready({
+    message: `Server running at: ${server.info.uri}`,
+    badge: true
   })
+}
+
+process.on('unhandledRejection', error => consola.error(error))
+
+start()