+++ template/frameworks/iview/pages/index.vue
@@ -1,5 +1,5 @@
 <template>
-  <section class="container">
+  <div class="container">
     <div>
       <logo />
       <h1 class="title">
@@ -9,12 +9,18 @@
         Welcome to the iView + Nuxt.js template
       </h2>
       <div class="links">
-        <Button type="primary" target="_blank" to="https://nuxtjs.org/">Documentation</Button>
-        <Button target="_blank" to="https://github.com/nuxt/nuxt.js">GitHub</Button>
-        <Button target="_blank" to="https://www.iviewui.com/">iView</Button>
+        <Button type="primary" target="_blank" to="https://nuxtjs.org/">
+          Documentation
+        </Button>
+        <Button target="_blank" to="https://github.com/nuxt/nuxt.js">
+          GitHub
+        </Button>
+        <Button target="_blank" to="https://www.iviewui.com/">
+          iView
+        </Button>
       </div>
     </div>
-  </section>
+  </div>
 </template>
 
 <script>