+++ CHANGELOG.md
@@ -2,6 +2,88 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [2.9.1](https://github.com/nuxt/create-nuxt-app/compare/v2.9.0...v2.9.1) (2019-08-04)
+
+
+### Bug Fixes
+
+* cannot find module babel-core ([#310](https://github.com/nuxt/create-nuxt-app/issues/310)) ([cf1dbde](https://github.com/nuxt/create-nuxt-app/commit/cf1dbde))
+* framevuerk error ([250c6e1](https://github.com/nuxt/create-nuxt-app/commit/250c6e1))
+* ignore vim swap files ([#263](https://github.com/nuxt/create-nuxt-app/issues/263)) ([2ff7b11](https://github.com/nuxt/create-nuxt-app/commit/2ff7b11))
+* remove unnecessary nodemon in devDependencies ([#309](https://github.com/nuxt/create-nuxt-app/issues/309)) ([cddccd6](https://github.com/nuxt/create-nuxt-app/commit/cddccd6))
+
+
+### Features
+
+* upgrade eslint-module ([#300](https://github.com/nuxt/create-nuxt-app/issues/300)) ([b61e88e](https://github.com/nuxt/create-nuxt-app/commit/b61e88e))
+
+
+
+# [2.9.0](https://github.com/nuxt/create-nuxt-app/compare/v2.8.0...v2.9.0) (2019-07-26)
+
+
+### Bug Fixes
+
+* Adonis commands always running nuxt build ([#268](https://github.com/nuxt/create-nuxt-app/issues/268)) ([d38aa96](https://github.com/nuxt/create-nuxt-app/commit/d38aa96))
+* Duplicate key 'devModules' ([#273](https://github.com/nuxt/create-nuxt-app/issues/273)) ([e7ef3a9](https://github.com/nuxt/create-nuxt-app/commit/e7ef3a9))
+* prettier error in vuetify ([#281](https://github.com/nuxt/create-nuxt-app/issues/281)) ([b7756a2](https://github.com/nuxt/create-nuxt-app/commit/b7756a2))
+
+
+### Features
+
+* add --verbose cli flag ([0d964e2](https://github.com/nuxt/create-nuxt-app/commit/0d964e2))
+* add alias for --edge and --info ([7d0522d](https://github.com/nuxt/create-nuxt-app/commit/7d0522d))
+* add UI framework Framevuerk ([#291](https://github.com/nuxt/create-nuxt-app/issues/291)) ([db8441e](https://github.com/nuxt/create-nuxt-app/commit/db8441e))
+* lint git staged files ([#269](https://github.com/nuxt/create-nuxt-app/issues/269)) ([1760f6a](https://github.com/nuxt/create-nuxt-app/commit/1760f6a))
+* replace tailwind with @nuxtjs/tailwindcss ([#254](https://github.com/nuxt/create-nuxt-app/issues/254)) ([5d0c6ff](https://github.com/nuxt/create-nuxt-app/commit/5d0c6ff))
+* upgrade Ava ([#284](https://github.com/nuxt/create-nuxt-app/issues/284)) ([54c29ff](https://github.com/nuxt/create-nuxt-app/commit/54c29ff))
+* upgrade vuetify to 2.x ([#297](https://github.com/nuxt/create-nuxt-app/issues/297)) ([f735b63](https://github.com/nuxt/create-nuxt-app/commit/f735b63))
+* **chore:** included a dedicated option to show up envinfo ([#272](https://github.com/nuxt/create-nuxt-app/issues/272)) ([d2917a9](https://github.com/nuxt/create-nuxt-app/commit/d2917a9))
+* **cli:** use cac for cli ([#264](https://github.com/nuxt/create-nuxt-app/issues/264)) ([b80fa2d](https://github.com/nuxt/create-nuxt-app/commit/b80fa2d))
+
+
+
+# [2.8.0](https://github.com/nuxt/create-nuxt-app/compare/v2.7.1...v2.8.0) (2019-06-24)
+
+
+### Bug Fixes
+
+* iview UI package version match ([#265](https://github.com/nuxt/create-nuxt-app/issues/265)) ([e547353](https://github.com/nuxt/create-nuxt-app/commit/e547353))
+
+
+### Features
+
+* move linting tool to separete prompt ([a6a4ac9](https://github.com/nuxt/create-nuxt-app/commit/a6a4ac9))
+
+
+
+## [2.7.1](https://github.com/nuxt/create-nuxt-app/compare/v2.7.0...v2.7.1) (2019-06-06)
+
+
+### Bug Fixes
+
+* undefined colors vuetify errors in nuxt.config.js ([650fe6c](https://github.com/nuxt/create-nuxt-app/commit/650fe6c))
+
+
+
+# [2.7.0](https://github.com/nuxt/create-nuxt-app/compare/v2.6.0...v2.7.0) (2019-06-06)
+
+
+### Bug Fixes
+
+* **antd-ui:** register plugin only once ([#233](https://github.com/nuxt/create-nuxt-app/issues/233)) ([afe1c6c](https://github.com/nuxt/create-nuxt-app/commit/afe1c6c))
+* **element:** register element UI only once ([#232](https://github.com/nuxt/create-nuxt-app/issues/232)) ([7e533dd](https://github.com/nuxt/create-nuxt-app/commit/7e533dd))
+* **iview:** register plugin only once ([#234](https://github.com/nuxt/create-nuxt-app/issues/234)) ([7ab985d](https://github.com/nuxt/create-nuxt-app/commit/7ab985d))
+
+
+### Features
+
+* **hapi:** upgrade to hapi 18 ([#231](https://github.com/nuxt/create-nuxt-app/issues/231)) ([3112b75](https://github.com/nuxt/create-nuxt-app/commit/3112b75))
+* **vuetify:** use vuetify module ([#225](https://github.com/nuxt/create-nuxt-app/issues/225)) ([1c8d944](https://github.com/nuxt/create-nuxt-app/commit/1c8d944))
+* upgrade tailwindcss to v1 ([#237](https://github.com/nuxt/create-nuxt-app/issues/237)) ([b429533](https://github.com/nuxt/create-nuxt-app/commit/b429533))
+
+
+
 # [2.6.0](https://github.com/nuxt/create-nuxt-app/compare/v2.5.1...v2.6.0) (2019-03-13)